Create FWRegexValidator class and refactor FWValidator
