+++ Sources/Formworks/FWComponents/FWRegexValidator.swift
@@ -1,0 +1,50 @@
+//
+//  FWRegexValidator.swift
+//  Formworks
+//
+//  Created by Edgar Sgroi on 29/10/20.
+//
+
+import Foundation
+
+enum FWDefaultRegexValidator {
+    case email
+    case phonenumber
+    case cellphone
+    case max32
+}
+
+final class FWRegexValidator: FWValidator {
+    let regexs: [String]
+    
+    init(regexs: [String]) {
+        self.regexs = regexs
+    }
+    
+    init(defaultValidator: FWDefaultRegexValidator) {
+        self.regexs = FWRegexValidator.getDefaultRegex(validator: defaultValidator)
+    }
+    
+    func validate(_ content: String) -> Bool {
+        for regex in regexs {
+            let predicate = NSPredicate(format: "SELF MATCHES %@", regex)
+            if predicate.evaluate(with: content) {
+                return true
+            }
+        }
+        return false
+    }
+    
+    static private func getDefaultRegex(validator: FWDefaultRegexValidator) -> [String] {
+        switch validator {
+        case .email:
+            return ["[0-9a-z._%+-]+@[a-z0-9.-]+\\.[a-z]{2,64}"]
+        case .phonenumber:
+            return ["[0-9]{2} [0-9]{8}", "[0-9]{2} [0-9]{4}-[0-9]{4}"]
+        case .cellphone:
+            return ["[0-9]{2} [0-9]{9}", "[0-9]{2} [0-9]{5}-[0-9]{4}"]
+        case .max32:
+            return ["[A-Za-z0-9 !\"#$%&'()*+,-./:;<=>?@\\[\\\\\\]^_`{|}~]{0,32}"]
+        }
+    }
+}
