+++ Sources/Formworks/FWComponents/FWValidator.swift
@@ -7,35 +7,7 @@
 
 import Foundation
 
-enum FWValidator {
-    case email
-    case phonenumber
-    case cellphone
-    case max32
-}
 
-extension FWValidator {
-    func regex() -> [String] {
-        switch self {
-        case .email:
-            return ["[0-9a-z._%+-]+@[a-z0-9.-]+\\.[a-z]{2,64}"]
-        case .phonenumber:
-            return ["[0-9]{2} [0-9]{8}", "[0-9]{2} [0-9]{4}-[0-9]{4}"]
-        case .cellphone:
-            return ["[0-9]{2} [0-9]{9}", "[0-9]{2} [0-9]{5}-[0-9]{4}"]
-        case .max32:
-            return ["[A-Za-z0-9 !\"#$%&'()*+,-./:;<=>?@\\[\\\\\\]^_`{|}~]{0,32}"]
-        }
-    }
-    
-    func validate(_ content: String) -> Bool {
-        let regexs = self.regex()
-        for regex in regexs {
-            let predicate = NSPredicate(format: "SELF MATCHES %@", regex)
-            if predicate.evaluate(with: content) {
-                return true
-            }
-        }
-        return false
-    }
+protocol FWValidator: AnyObject {
+    func validate(_ content: String) -> Bool
 }
